@@ -1,13 +1,9 @@
 package by.dz.fruits_in_choco.fruits_in_choco.dto;
 
-import lombok.Builder;
-import lombok.Data;
-
 import java.util.Date;
 
-@Data
-@Builder
-public class Notification {
-    private Date date;
-    private String type;
+public record Notification(
+        Date date,
+        String type
+) {
 }
