@@ -1,12 +1,10 @@
 package by.dz.fruits_in_choco.fruits_in_choco.dto;
 
-import lombok.Data;
-
-@Data
-public class ProductRatingRequest {
-    private Long userId;
-    private String author;
-    private int rating;
-    private String message;
-    private boolean approved;
+public record ProductRatingRequest(
+        Long userId,
+        String author,
+        int rating,
+        String message,
+        boolean approved
+) {
 }
