@@ -1,11 +1,9 @@
 package by.dz.fruits_in_choco.fruits_in_choco.dto;
 
-import lombok.Data;
-
-@Data
-public class RegistrationRequest {
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
+public record RegistrationRequest(
+        String firstName,
+        String lastName,
+        String email,
+        String password
+) {
 }
