@@ -1,9 +1,6 @@
 package by.dz.fruits_in_choco.fruits_in_choco.dto.auth;
 
-import lombok.Data;
-
-@Data
-public class AuthenticationRequest {
-    private String email;
-    private String password;
-}
+public record AuthenticationRequest(
+        String email,
+        String password
+) { }
